App refactor: Introduce StdoutContextProvider
Improve App module's read- and compose-ability by moving StdoutContext
concerns to a separate provider module.
